Create classi DataLayer per metodi SchoolPeriods e EraseClasses
