+++ DataLayer/EraseClasses.cs
@@ -1,0 +1,49 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.DataLayer
+{
+    class EraseClasses
+    {
+
+        DataLayer dl = new DataLayer();
+
+        internal void EraseClassFromClasses(Class Class)
+        {
+            //EraseAllStudentsOfAClass(Class); 
+            using (DbConnection conn = dl.Connect())
+            {
+                // delete all the references in link table between students and classes
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Classes_Students" +
+                    " WHERE Classes_Students.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from Classes_SchoolSubjects
+                cmd.CommandText = "DELETE FROM Classes_SchoolSubjects" +
+                    " WHERE Classes_SchoolSubjects.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from Classes_Tests
+                cmd.CommandText = "DELETE FROM Classes_Tests" +
+                    " WHERE Classes_Tests.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // erase class from table Classes 
+                cmd.CommandText = "DELETE FROM Classes" +
+                    " WHERE Classes.idClass=" + Class.IdClass +
+                    ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+    }
+}
