+++ DbClasses/User.cs
@@ -6,6 +6,7 @@ namespace SchoolGrades.DbClasses
 {
     class User
     {
+        //Campi di classe
         string username;
         string firstName;
         string lastName;
@@ -16,6 +17,7 @@ namespace SchoolGrades.DbClasses
         string salt;
         int? idUserCategory;
 
+        //Proprietà per ogni campo di classe
         public string Username { get => username; set => username = value; }
         public string LastName { get => lastName; set => lastName = value; }
         public string Password { get => password; set => password = value; }
@@ -29,18 +31,27 @@ namespace SchoolGrades.DbClasses
         public int? IdUserCategory { get => idUserCategory; set => idUserCategory = value; }
         public bool? IsEnabled { get; internal set; }
 
+        /// <summary>
+        /// Metodo costruttore di un Utente base con l'inserimento di un username e di una password
+        /// </summary>
+        /// <param name="Username">Username dell'utente</param>
+        /// <param name="Password">Password dell'utente</param>
         public User(string Username, string Password)
         {
+            //Nel caso l'utente non abbia nessuna username e password viene creata una eccezione
             if ((Username == null) || (Password == null))
                 throw new Exception("Username o password non accettabile");
+            //Sostituzione del valore del campo di classe con il nuovo valore
             this.username = Username;
             this.password = Password;
         }
 
         public User(string Username, string Password, string Description, string LastName, string FirstName, string Email)
         {
+            //Nel caso l'utente non abbia una delle informazioni necessarie per la creazione viene inviata una nuova eccezione
             if ((Username == null) || (Password == null) || (Description == null) || (LastName == null) || (FirstName == null) || (Email == null))
                 throw new Exception("Informazioni ricevute non accettabili");
+            //Sostituzione del valore del campo di classe con il nuovo valore
             this.username = Username;
             this.password = Password;
             this.description = Description;
