+++ SchoolGrades/frmUsersManagement.cs
@@ -1,0 +1,58 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Text;
+using System.Windows.Forms;
+
+namespace SchoolGrades
+{
+    public partial class frmUsersManagement : Form
+    {
+        public frmUsersManagement()
+        {
+            InitializeComponent();
+        }
+
+        /// <summary>
+        /// Save the new password of the user
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void savePassword_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        /// <summary>
+        /// Change the old password of the user with the new one from the textbox
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void changePassword_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        /// <summary>
+        /// Creates a new user with the information from the textboxs
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void newUser_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        /// <summary>
+        /// Save the information in the textbox
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void saveInformation_Click(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
